refactor: Move helper funcs to utils.py
